requirements.txt
streamlit==1.32.2
Pillow==9.5.0
numpy==1.24.4
imageio==2.34.0
